% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{add_negated_powers}
\alias{add_negated_powers}
\title{Create a sequence of negative powers on a closed interval}
\usage{
add_negated_powers(lo, hi, mult = 8)
}
\arguments{
\item{lo}{bottom of the interval}

\item{hi}{top of the interval}

\item{mult}{multiplier}
}
\value{
an \code{integer} vector
}
\description{
Calls \link{add_powers} to produce sequences of negative powers. This
function will only produce sequences of negative integers.
}
\note{
This is not intended to be called directly by users, but it is exported
in case it is useful to be imported into another package.
}
\examples{
add_negated_powers(lo = -bitwShiftL(a = 8, n = 10), hi = -8, mult = 8)
}
