% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_negated_powers_cpp}
\alias{add_negated_powers_cpp}
\title{add negated powers cpp}
\usage{
add_negated_powers_cpp(lo, hi, mult)
}
\description{
add negated powers cpp
}
\examples{
add_negated_powers_cpp(lo = -bitwShiftL(a = 8, n = 10), hi = 0, mult = 8)
}
