% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{add_powers}
\alias{add_powers}
\title{Append the powers of \code{mult} in the closed interval [\code{lo}, \code{hi}]}
\usage{
add_powers(lo, hi, mult = 8)
}
\arguments{
\item{lo}{bottom of the interval}

\item{hi}{top of the interval}

\item{mult}{multiplier}
}
\value{
an \code{integer} vector
}
\description{
Generate a sequence consisting of the powers of \code{mult} on the closed interval [\code{lo}, \code{hi}].
This function works on integer sequences, so if an interval is requested outside of
the maximum representable range of integers \eqn{[-(2^{31} - 1), 2^{31} - 1]} it
will throw an error. This function will only produce sequences of positive integers.
}
\examples{
add_powers(lo = 8, hi = bitwShiftL(a = 8, n = 10))
}
