% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_powers_cpp}
\alias{add_powers_cpp}
\title{add powers cpp}
\usage{
add_powers_cpp(lo, hi, mult)
}
\description{
add powers cpp
}
\examples{
add_powers_cpp(lo = 8, hi = bitwShiftL(a = 8, n = 10), mult = 8)
}
