% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{add_range}
\alias{add_range}
\title{Create a logarithmic range of integers}
\usage{
add_range(lo, hi, mult = 8, exclude_zero = FALSE)
}
\arguments{
\item{lo}{bottom of the interval}

\item{hi}{top of the interval}

\item{mult}{multiplier}

\item{exclude_zero}{include \code{0} in the interval or not}
}
\description{
Create a range from the powers of \code{mult} on the closed interval [\code{lo}, \code{hi}].
The endpoints are included, even if they are not multiples of \code{mult}. The range
may go from negative to positive values if requested.
}
\examples{
add_range(lo = 8, hi = bitwShiftL(a = 8, n = 10), mult = 2)
}
