% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_range_cpp}
\alias{add_range_cpp}
\title{create a range}
\usage{
add_range_cpp(lo, hi, mult)
}
\description{
create a range
}
\examples{
add_range_cpp(lo = -bitwShiftL(a = 8, n = 10), hi = bitwShiftL(a = 8, n = 10), mult = 8)
}
